+++ test/screenwise_test.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:flutter_adaptive_layouts/flutter_adaptive_layouts.dart';
+import 'package:screenwise/screenwise.dart';
 
 void main() {
   testWidgets('ResponsiveRow builds correctly', (WidgetTester tester) async {