+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 0.0.1
 
-* Initial release of Flutter Adaptive Layouts package.
+* Initial release of ScreenWise package.
 * Added responsive widgets that adapt to different screen sizes.
 * Included ResponsiveRow, ResponsiveGrid, AdaptiveColumn, ResponsiveBuilder, ResponsiveContainer, ResponsiveScaffold, ResponsiveText, and ResponsiveValue.
 * Added utilities for screen size detection and responsive configurations.